Added comments and tooltips to all PlayerAttachable types, did some refactoring. - Corrected some erroneous and/or outdated tooltips. - Added some fields to configure camera clamping for PlayerSeater.   - Tooltips for hidden items are now properly hidden, too. - Made it so if a player action is not found for whatever reason, it'll throw a warning and disable the script instead of spamming errors. - Made range parameter for SpraySensor and FearInducer a floating point number instead of an integer.

+++ itolib/Behaviours/Effects/FearInducer.cs
@@ -33,8 +33,8 @@ namespace itolib.Behaviours.Effects
         ///     TODO.
         /// </summary>
         [Tooltip("")]
-        [Min(0)]
-        [SerializeField] private int lookRange = 60;
+        [Min(0.0f)]
+        [SerializeField] private float lookRange = 60.0f;
 
         /// <summary>
         ///     TODO.
